auth_service/requirements.txt
alembic==1.6.5
click==8.0.1
Flask==2.0.1
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
flask-jwt-extended==4.2.3
flask-marshmallow==0.14.0
greenlet==1.1.0
install==1.3.4
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
redis==3.5.3
six==1.16.0
SQLAlchemy==1.4.20
Werkzeug==2.0.1
psycopg2-binary==2.9.1
sqlalchemy-utils==0.37.8
gevent==21.1.2
gunicorn==20.1.0
alembic==1.6.5
click==8.0.1
dnspython==2.1.0
Flask==2.0.1
flask-marshmallow==0.14.0
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
flask_jwt_extended==4.2.3
gevent==21.1.2
greenlet==1.1.0
gunicorn==20.1.0
idna==3.2
install==1.3.4
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
marshmallow==3.12.2
marshmallow-sqlalchemy==0.26.1
marshmallow-validators==3.0.0
psycopg2-binary==2.9.1
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
redis==3.5.3
six==1.16.0
SQLAlchemy==1.4.20
SQLAlchemy-Utils==0.37.8
Werkzeug==2.0.1
zope.event==4.5.0
zope.interface==5.4.0
